Implement stack-preserving THROW/CATCH
Stack-preserving THROW/CATCH are like the ANS ones, except THROW does
not inspect or affect the stack at all.  This means that the data stack
depth after an exceptional return from CATCH cannot be known statically.
For now, a normal return from CATCH still pushes a zero on the data
stack, but this is subject to change.